Removed domain restrictions on logicInput, logicOutput and logicParameter object properties
Removed the domain restriction of logicInput, logicOutput and logicParameter object properties

The main aim is to relax semantic restrictions here when referring via the described properties to entities within a custom model of a control logic. The original domain restriction imply the classification of each related object as a member of the class ApplicationLogic. This might cause problems and hence, the domain restriction is removed here.